refactor(sql): extract shared AbstractDenseScanAsOfJoinRecordCursor
Move the resilient forward+backward Dense keyed-ASOF scan (maps, scan state,
resolveViaDenseScan, scanForward, key hooks) out of AsOfJoinDenseRecordCursorBase
into a new intermediate class both the Dense cursor and (next) the Memoized cursor
can drive. Behaviour-preserving: Dense hasNext now delegates to resolveViaDenseScan;
AsOfJoinTest 120/120 + AsOfJoinFuzzTest 6/6 green.
